Deal with this case:
Create .#core in a directory, and create a symlink called "core" pointing
to /usr/tmp/core.  Then run "expunge core" and you'll get an error and th
the expunge will exit without getting rid of .#core.  It will now print an
error about the symlink, but it will expunge the .#core anyway.